Cleanup cvs dump module
